+++ test-support/learner-it-support/src/main/java/de/learnlib/testsupport/it/learner/PassiveLearnerVariantListImpl.java
@@ -21,6 +21,7 @@ import java.util.List;
 import de.learnlib.api.algorithm.PassiveLearningAlgorithm;
 import net.automatalib.automaton.fsa.DFA;
 import net.automatalib.automaton.transducer.MealyMachine;
+import net.automatalib.automaton.transducer.MooreMachine;
 import net.automatalib.automaton.transducer.SubsequentialTransducer;
 import net.automatalib.word.Word;
 
@@ -43,6 +44,10 @@ public class PassiveLearnerVariantListImpl<M, I, D> implements PassiveLearnerVar
     public static class MealyLearnerVariantListImpl<I, O>
             extends PassiveLearnerVariantListImpl<MealyMachine<?, I, ?, O>, I, Word<O>>
             implements MealyLearnerVariantList<I, O> {}
+
+    public static class MooreLearnerVariantListImpl<I, O>
+            extends PassiveLearnerVariantListImpl<MooreMachine<?, I, ?, O>, I, Word<O>>
+            implements MooreLearnerVariantList<I, O> {}
 
     public static class SSTLearnerVariantListImpl<I, O>
             extends PassiveLearnerVariantListImpl<SubsequentialTransducer<?, I, ?, O>, I, Word<O>>